requirements.txt
opencv-python
mss
pytesseract
fuzzywuzzy
pyyaml
